+++ retro/reco.py
@@ -60,7 +60,8 @@ from retro.utils.misc import expand, mkdir, sort_dict
 from retro.utils.stats import estimate_from_llhp
 from retro.priors import get_prior_fun
 from retro.hypo.discrete_muon_kernels import pegleg_eval
-from retro.tables.pexp_5d import generate_pexp_and_llh_functions
+from retro.llh import generate_llh_function
+from retro.pexp import generate_pexp_function
 from retro.hypo.discrete_cascade_kernels import SCALING_CASCADE_ENERGY
 
 
@@ -154,7 +155,13 @@ class Reco(object):
         mkdir(self.outdir)
         self.dom_tables = init_obj.setup_dom_tables(**dom_tables_kw)
         self.tdi_tables, self.tdi_metas = init_obj.setup_tdi_tables(**tdi_tables_kw)
-        self.pexp, self.get_llh, _ = generate_pexp_and_llh_functions(
+        self.pexp_, self.pexp, _ = generate_pexp_function(
+            dom_tables=self.dom_tables,
+            tdi_tables=self.tdi_tables,
+            tdi_metas=self.tdi_metas,
+        )
+        self.get_llh = generate_llh_function(
+            pexp=self.pexp_,
             dom_tables=self.dom_tables,
             tdi_tables=self.tdi_tables,
             tdi_metas=self.tdi_metas,
@@ -853,6 +860,7 @@ class Reco(object):
                 generic_sources=generic_sources,
                 pegleg_sources=pegleg_sources,
                 scaling_sources=scaling_sources,
+                scaling_cascade_energy=scaling_cascade_energy,
                 event_hit_info=event_hit_info,
                 event_dom_info=event_dom_info,
                 pegleg_stepsize=1,
